# openapi/openapi-helpdesk.json
{
  "openapi" : "3.0.3",
  "components" : {
    "schemas" : {
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string",
            "example" : "Internal Server Error"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer",
            "example" : 500
          },
          "detail" : {
            "type" : "string",
            "example" : "An unexpected error has occurred. Please contact support."
          },
          "instance" : {
            "type" : "string",
            "example" : "PDFS-500"
          }
        }
      },
      "InfoResponse" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "pagopa-receipt-pdf-service"
          },
          "version" : {
            "type" : "string",
            "example" : "1.2.3"
          },
          "environment" : {
            "type" : "string",
            "example" : "dev"
          },
          "description" : {
            "type" : "string",
            "example" : "Receipt PDF Service"
          }
        }
      }
    },
    "responses" : {
      "AppException400" : {
        "description" : "Default app exception for status 400",
        "content" : {
          "application/json" : {
            "examples" : {
              "Error" : {
                "value" : {
                  "type" : "",
                  "title" : "Bad Request",
                  "status" : 400,
                  "detail" : "The provided third party id [<td_id>] is invalid",
                  "instance" : "PDFS_703"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "AppException404" : {
        "description" : "Default app exception for status 404",
        "content" : {
          "application/json" : {
            "example" : {
              "type" : "",
              "title" : "Not Found",
              "status" : 404,
              "detail" : "Third party id [<td_id>] not found",
              "instance" : "PDFS_900"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError" : {
        "description" : "Internal Server Error",
        "content" : {
          "application/json" : {
            "example" : {
              "type" : "",
              "title" : "Internal Server Error",
              "status" : 500,
              "detail" : "An unexpected error has occurred. Please contact support.",
              "instance" : "PDFS_603"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "info" : {
    "title" : "PagoPa PDF Receipt Service (local)",
    "version" : "0.0.0-SNAPSHOT",
    "description" : "Expose APIs used by IO backend to retrieved payment receipts",
    "termsOfService" : "https://www.pagopa.gov.it/"
  },
  "tags" : [ {
    "name" : "Helpdesk",
    "description" : "Helpdesk operations"
  }, {
    "name" : "Info",
    "description" : "Info operations"
  } ],
  "paths" : {
    "/helpdesk/cart-errors-toreview/{cart-id}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "cart-id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Cart Receipt Error By Cart Id"
      }
    },
    "/helpdesk/cart-receipts/io-message/{message-id}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "message-id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Cart Receipt Message"
      }
    },
    "/helpdesk/cart-receipts/organizations/{organization-fiscal-code}/iuvs/{iuv}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "iuv",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "organization-fiscal-code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Cart Receipt By Organization Fiscal Code And IUV"
      }
    },
    "/helpdesk/cart-receipts/{cart-id}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "cart-id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Cart Receipt"
      }
    },
    "/helpdesk/errors-toreview/{bizevent-id}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "bizevent-id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Receipt Error By Event Id"
      }
    },
    "/helpdesk/pdf-receipts/{file-name}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "file-name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Receipt Pdf"
      }
    },
    "/helpdesk/receipts/io-message/{message-id}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "message-id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Receipt Message"
      }
    },
    "/helpdesk/receipts/organizations/{organization-fiscal-code}/iuvs/{iuv}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "iuv",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "organization-fiscal-code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Receipt By Organization Fiscal Code And IUV"
      }
    },
    "/helpdesk/receipts/{event-id}" : {
      "get" : {
        "tags" : [ "Helpdesk" ],
        "parameters" : [ {
          "name" : "event-id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Receipt"
      }
    },
    "/info" : {
      "get" : {
        "summary" : "Get info of Receipt PDF Service",
        "tags" : [ "Info" ],
        "responses" : {
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InfoResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}